HDFS-6125. Cleanup unnecessary cast in HDFS code base. Contributed by Suresh Srinivas.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/trunk@1581242 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/web/JsonUtil.java
@@ -317,7 +317,7 @@ public class JsonUtil {
     if (value == null) {
       return defaultValue;
     }
-    return (long) (Long) value;
+    return (Long) value;
   }
 
   private static String getString(Map<?, ?> m, String key,
